pyproject.toml
[tool.poetry]
name = "pydanticAi-langchain"
version = "0.1.0"
description = "Multi-agent LLM system using LangGraph and PydanticAI"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "*" }]

[tool.poetry.dependencies]
python = "^3.9.10"
pydantic-ai = "0.2.15"
langgraph = "0.4.8"
openai = "1.84.0"
logfire = "^3.18.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
streamlit = "^1.37.0"
pillow = "^10.4.0"
requests-oauthlib = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"   
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
llm-agents = "main:main"